sections/cesar/schmidt_rank_project/requirements.txt
# Django
Django==5.0.1

# Mathematics and Scientific Computing
numpy==1.26.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
Pillow==10.2.0

# Forms
django-crispy-forms>=2.3
crispy-bootstrap5==2025.6

# Utilities
python-dateutil==2.8.2
